bunq-api/pyproject.toml
[project]
name = "bunq-api"
version = "0.1.0"
description = "bunq Hackathon API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "requests",
    "cryptography",
    "bunq-sdk>=1.28.0",
    "flask-sqlalchemy>=3.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bunq_api"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.4.0",
]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
